added square

+++ Engine/src/main.cpp
@@ -1,9 +1,9 @@
 #include <Engine/Engine.hpp>
 
 int main() {
-    GLFWwindow* window;
-    Window      WindowSettings;
-    scripts::MainScript  Script;
+    GLFWwindow *window;
+    Window WindowSettings;
+    scripts::MainScript Script;
 
     Script.Start();
 
@@ -29,14 +29,14 @@ int main() {
     }
 
 #ifndef NDEBUG
-    std::cout << "OpenGL " << glGetString(GL_VERSION)  << std::endl;
-    std::cout << "GPU "    << glGetString(GL_RENDERER) << std::endl;
+    std::cout << "OpenGL " << glGetString(GL_VERSION) << std::endl;
+    std::cout << "GPU " << glGetString(GL_RENDERER) << std::endl;
 #endif // NDEBUG
 
     Color bc = WindowSettings.get_background_color();
     glClearColor(bc.r / 255.0, bc.g / 255.0, bc.b / 255.0, bc.a);
 
-    glScalef(ws.y / (float)ws.x, 1, 1);
+    glScalef(ws.y / ws.x, 1, 1);
 
     while (!glfwWindowShouldClose(window)) {
         glClear(GL_COLOR_BUFFER_BIT);
